[CSR256HW-5]: implement controllers

+++ src/PriceCalculator.Api/Responses/V1/GetHistoryResponse.cs
@@ -6,8 +6,8 @@ public record GetHistoryResponse(
 )
 {
     public record CargoResponse(
-        decimal Volume,
-        decimal Weight,
+        double Volume,
+        double Weight,
         long[] GoodIds
     );
 }
